backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
sqlalchemy>=2.0,<3.0
psycopg[binary]>=3.2,<4.0
pydantic-settings>=2.7,<3.0
alembic>=1.13,<2.0
passlib[bcrypt]>=1.7,<2.0
python-jose[cryptography]>=3.3,<4.0

# Testing
pytest>=8.0,<9.0
httpx>=0.27,<1.0

# Linting
ruff>=0.8,<1.0
